backend/routes/inspiration.py
"""
Inspiration Routes
API endpoints for AI-powered inspiration generation
"""

from flask import Blueprint, request, jsonify
import google.generativeai as genai

bp = Blueprint('inspiration', __name__)

@bp.route('/generate', methods=['POST'])
def generate_inspiration():
    """Generate inspiration from input (text, image, audio)"""
    data = request.json
    input_type = data.get('type', 'text')
    content = data.get('content', '')
    
    # Placeholder for inspiration generation
    return jsonify({
        'success': True,
        'ideas': [
            'What if the protagonist discovers a hidden ability?',
            'Consider exploring the antagonist\'s backstory',
            'A twist: the mentor figure has been manipulating events'
        ],
        'prompts': [
            'Write a scene where the truth is revealed',
            'Explore the emotional aftermath of this discovery'
        ]
    })

@bp.route('/scenarios', methods=['POST'])
def generate_scenarios():
    """Generate 'what if' scenarios"""
    data = request.json
    context = data.get('context', '')
    
    return jsonify({
        'success': True,
        'scenarios': [
            'What if the main character made the opposite choice?',
            'What if a trusted ally is secretly working against them?',
            'What if the setting changed to a different time period?'
        ]
    })
