backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
firebase-admin==6.4.0
gunicorn==21.2.0
pytest==7.4.3
requests==2.31.0
pillow==10.1.0
pydub==0.25.1
pypdf>=4.0.0
ebooklib==0.18
markdown==3.5.1
